refactor(Referee): Rule -> Referee와 통합
